리팩터 : api 연결 함수 분리 - 기기 위치 및 정보 서버 연결하기 함수와 기관의 포인트 랭킹 조화 함수 파일 분리
